#1161 Update web sample to use declarative router
